refactor(CheckList):修改文件名称，删除多余文件
